requirements.txt
BeautifulSoup==3.2.1
blessings==1.6
charm-tools==2.1.4
Cheetah==2.4.4
colander==1.3.1
ecdsa==0.13
functools32==3.2.3.post2
httplib2==0.9.2
iso8601==0.1.11
jsonschema==2.5.1
jujubundlelib==0.5.2
keyring==9.3.1
launchpadlib==1.10.4
lazr.authentication==0.1.3
lazr.restfulclient==0.13.1
lazr.uri==1.0.3
libcharmstore==0.0.3
lxml==3.6.4
Markdown==2.6.6
oauth==1.0.1
otherstuf==1.1.0
paramiko==1.17.2
parse==1.6.6
path.py==8.2.1
pathspec==0.5.0
pbr==1.10.0
pkg-resources==0.0.0
pycrypto==2.6.1
PyYAML==3.12
requests==2.11.1
ruamel.ordereddict==0.4.9
ruamel.yaml==0.12.10
SecretStorage==2.2.1
simplejson==3.8.2
six==1.10.0
stuf==0.9.16
testresources==2.0.1
theblues==0.3.4
translationstring==1.3
typing==3.5.2.2
virtualenv==15.0.3
wadllib==1.3.2
wsgi-intercept==1.3.1
xlrd==1.0.0
zc.buildout==2.5.3
zope.interface==4.3.2
